feat: API tests
This implements tests for the FastAPI application.
There's two supported modes:
- Run tests directly against the app, as described in https://fastapi.tiangolo.com/tutorial/testing/
- Run tests by performing HTTP requests against a server,
  e.g. for containerized e2e testing.

To run tests against a server, set the `BYSYKKEL_E2E_TEST_HOST` env
variable to the base URL of the deployment.
For local testing, set it to `http://localhost:8000`

If the env variable is not set, tests are run directly against the
FastAPI ASGI app.